refactor class and props name
